refactor: move out new window to a separate file
