+++ job/window_v1.go
@@ -9,21 +9,6 @@ import (
 	"github.com/odpf/optimus/utils"
 )
 
-type Window interface {
-	GetTimeRange(scheduleTime time.Time) (time.Time, time.Time, error)
-	Validate() error
-}
-
-func GetWindow(version int, truncateTo, offset, size string) Window {
-	if version == 1 {
-		return WindowV1{size, offset, truncateTo}
-	}
-	if version == 2 {
-		return WindowV2{size, offset, truncateTo}
-	}
-	return nil
-}
-
 type WindowV1 struct {
 	Size       string
 	Offset     string
